+++ src/main/kotlin/data/audit/mapper/AuditMapper.kt
@@ -14,15 +14,6 @@ import java.time.LocalDateTime
 import java.util.*
 
 class AuditMapper {
-    /*companion object {
-        const val ID_FIELD = "id"
-        const val CREATED_BY_FIELD = "createdBy"
-        const val ENTITY_ID_FIELD = "entityId"
-        const val ENTITY_TYPE_FIELD = "entityType"
-        const val OLD_STATE_FIELD = "oldState"
-        const val NEW_STATE_FIELD = "newState"
-        const val DATE_FIELD = "date"
-    }*/
 
     fun auditToDocument(audit: Audit): Document {
         return Document().apply {
