+++ src/test/kotlin/logic/project/EditStateOfProjectUIOfProjectUseCaseTest.kt
@@ -2,14 +2,14 @@ package logic.project
 
 import io.mockk.*
 import kotlinx.coroutines.test.runTest
+import logic.exceptions.ProjectNotFoundException
+import logic.exceptions.ProjectStateNotFoundException
+import logic.model.Project
+import logic.model.State
+import logic.repository.ProjectRepository
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import squad.abudhabi.logic.exceptions.ProjectNotFoundException
-import squad.abudhabi.logic.exceptions.ProjectStateNotFoundException
-import squad.abudhabi.logic.model.Project
-import squad.abudhabi.logic.model.State
-import squad.abudhabi.logic.repository.ProjectRepository
 import java.util.*
 
 class EditStateOfProjectUIOfProjectUseCaseTest {
