+++ src/test/kotlin/logic/task/GetTaskByIdUseCaseTest.kt
@@ -5,11 +5,11 @@ import helper.createTask
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
+import logic.exceptions.TaskNotFoundException
+import logic.repository.TaskRepository
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import squad.abudhabi.logic.exceptions.TaskNotFoundException
-import squad.abudhabi.logic.repository.TaskRepository
 import java.util.*
 
 class GetTaskByIdUseCaseTest {
